+++ src/page__/index.test.ts
@@ -1,26 +1,26 @@
 import { ctx_ } from '@ctx-core/object'
 import { test } from 'uvu'
 import { equal, is } from 'uvu/assert'
-import { page__, page__error__, page__host__, page__params__, page__path__, page__query__ } from '../index.js'
+import { page$__set, page__error_, page__host_, page__params_, page__path_, page__query_ } from '../index.js'
 test('page__', ()=>{
 	const ctx = ctx_()
-	equal(page__host__(ctx).$, null)
-	equal(page__path__(ctx).$, null)
-	equal(page__params__(ctx).$, null)
-	equal(page__query__(ctx).$, null)
-	equal(page__error__(ctx).$, null)
+	equal(page__host_(ctx), null)
+	equal(page__path_(ctx), null)
+	equal(page__params_(ctx), null)
+	equal(page__query_(ctx), null)
+	equal(page__error_(ctx), null)
 	const error = new Error('test error')
-	page__(ctx).$ = {
+	page$__set(ctx, {
 		host: 'test-host',
 		path: '/test/path',
 		params: { test: 'params' },
 		query: { test: 'query' },
 		error,
-	}
-	equal(page__host__(ctx).$, 'test-host')
-	equal(page__path__(ctx).$, '/test/path')
-	equal(page__params__(ctx).$, { test: 'params' })
-	equal(page__query__(ctx).$, { test: 'query' })
-	is(page__error__(ctx).$, error)
+	})
+	equal(page__host_(ctx), 'test-host')
+	equal(page__path_(ctx), '/test/path')
+	equal(page__params_(ctx), { test: 'params' })
+	equal(page__query_(ctx), { test: 'query' })
+	is(page__error_(ctx), error)
 })
 test.run()