+++ src/sapper_preloading__/index.d.ts
@@ -1,4 +1,12 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { be__return_T } from '@ctx-core/object'
-export declare const sapper_preloading__:be__return_T<WritableAtom_<object|undefined>>
-export { sapper_preloading__ as sapper_preloading$_ }
+import type { Writable_ } from '@ctx-core/svelte'
+export declare const [
+	sapper_preloading$_,
+	sapper_preloading_,
+	sapper_preloading$__set,
+] = be_writable_triple_T<
+	Writable_<object>
+>
+export {
+	sapper_preloading$_ as sapper_preloading__,
+	sapper_preloading$_ as sapper_preloading$_,
+}